+++ src/utils/index.ts
@@ -1,4 +1,5 @@
 export * from './packageReader';
 export * from './npmClient';
+export * from './packageManager';
 export * from './versionUtils';
 export * from './notify';